+++ src/components/paths/line.ts
@@ -1,8 +1,8 @@
 import {
   useRenderAfterChildren,
   useRenderBeforeChildren,
-} from '../hooks/use-render.ts';
-import type { CXComponent, PropsWithChildren } from '../types.ts';
+} from '../../hooks/use-render.ts';
+import type { CXComponent, PropsWithChildren } from '../../types.ts';
 
 export type LineProps = PropsWithChildren<{
   startX: number;
