+++ src/index.ts
@@ -8,12 +8,12 @@ export * from './hooks/use-loop.ts';
 export * from './hooks/use-on.ts';
 export * from './hooks/use-window-size.ts';
 export * from './components/canvas/index.ts';
-export * from './components/rectangle.ts';
-export * from './components/ellipse.ts';
-export * from './components/circle.tsx';
-export * from './components/line.ts';
-export * from './components/path.ts';
-export * from './components/point.ts';
-export * from './components/arc-to.ts';
-export * from './components/text.ts';
+export * from './components/shapes/rectangle.ts';
+export * from './components/shapes/ellipse.ts';
+export * from './components/shapes/circle.tsx';
+export * from './components/paths/line.ts';
+export * from './components/paths/path.ts';
+export * from './components/paths/point.ts';
+export * from './components/paths/arc-to.ts';
+export * from './components/text/text.ts';
 export * from './create-context.ts';
